webapi/tct-webaudio-w3c-tests/webaudio/audiocontext_createIIRFilter.html

<!DOCTYPE html>
<!--
Copyright (c) 2016 Intel Corporation.

Redistribution and use in source and binary forms, with or without modification,
are permitted provided that the following conditions are met:

* Redistributions of works must retain the original copyright notice, this list
  of conditions and the following disclaimer.
* Redistributions in binary form must reproduce the original copyright notice,
  this list of conditions and the following disclaimer in the documentation
  and/or other materials provided with the distribution.
* Neither the name of Intel Corporation nor the names of its contributors
  may be used to endorse or promote products derived from this work without
  specific prior written permission.

THIS SOFTWARE IS PROVIDED BY INTEL CORPORATION "AS IS"
AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
ARE DISCLAIMED. IN NO EVENT SHALL INTEL CORPORATION BE LIABLE FOR ANY DIRECT,
INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY
OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
-->

<meta charset="utf-8">
<title>WebAudio Test: IIRFilterNode interface</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<link rel="author" title="Chunyan Wang" href="mailto:chunyanx.wang@intel.com">
<link rel="help" href="http://webaudio.github.io/web-audio-api/#idl-def-IIRFilterNode">
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
test(function() {
  var context = new AudioContext();
  var iIRFilterNode = context.createIIRFilter([1], [1]);
  assert_true(iIRFilterNode instanceof IIRFilterNode, "createIIRFilter method return IIRFilterNode object");
  assert_true("getFrequencyResponse" in iIRFilterNode, "IIRFilterNode has implemented getFrequencyResponse");
  assert_equals(iIRFilterNode.numberOfInputs, 1, "iIRFilterNode.numberOfInputs equals 1");
  assert_equals(iIRFilterNode.numberOfOutputs, 1, "iIRFilterNode.numberOfOutputs equals 1");
  assert_equals(iIRFilterNode.channelInterpretation, "speakers", "iIRFilterNode.channelInterpretation equals speakers");
  assert_equals(iIRFilterNode.channelCountMode, "max", "iIRFilterNode.channelCountMode equals max");
  var fz = new Float32Array(128);
  var mResponse = new Float32Array(128);
  var pResponse = new Float32Array(128);
  iIRFilterNode.getFrequencyResponse(fz, mResponse, pResponse);
}, "Check that the default values of IIRFilterNode's members as specified.");

test(function() {
  var context = new AudioContext();
  var iIRFilterNode = context.createIIRFilter([1], [1]);
  var fz = new Float32Array(128);
  var mResponse = new Float32Array(128);
  var pResponse = new Float32Array(64);
  assert_throws("NotSupportedError", function() {
    iIRFilterNode.getFrequencyResponse(fz, mResponse, pResponse);
  });
}, "Check that the A NotSupportedError should be thrown when phaseResponse is shorter than frequencyHz.");

test(function() {
  var context = new AudioContext();
  var iIRFilterNode = context.createIIRFilter([1], [1]);
  var fz = new Float32Array(128);
  var mResponse = new Float32Array(64);
  var pResponse = new Float32Array(128);
  assert_throws("NotSupportedError", function() {
    iIRFilterNode.getFrequencyResponse(fz, mResponse, pResponse);
  });
}, "Check that the A NotSupportedError should be thrown when magResponse is shorter than frequencyHz.");
</script>
